// frontend/src/api/models/GridIntersection.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type GridIntersection = {
    image: string;
    polyline_x: Array<number>;
    polyline_y: Array<number>;
    x_min: number;
    x_max: number;
    y_min: number;
    y_max: number;
};